Lots and lots of work

+++ grimoire2/src/standalone/ingredient.rs
@@ -1,17 +1,18 @@
-use crate::modifiermap::ModifierMap;
-use crate::theoretical::Theoretical;
-
+use crate::{modifiermap::ModifierMap, theoretical::Theoretical};
 
 #[derive(Default, Debug, Clone)]
 pub struct StandaloneIngredient {
-    pub weight: usize,
+    pub weight: u8,
     pub lore_multiplier: Theoretical<f64>,
     pub modifiers: ModifierMap,
 }
 
-
 impl StandaloneIngredient {
-    pub fn new(weight: usize, lore_multiplier: Theoretical<f64>, modifiers: ModifierMap) -> Self {
-        Self { weight, lore_multiplier, modifiers }
+    pub fn new(weight: u8, lore_multiplier: Theoretical<f64>, modifiers: ModifierMap) -> Self {
+        Self {
+            weight,
+            lore_multiplier,
+            modifiers,
+        }
     }
 }
